Improve structure for src/CIAClassificationApp.tsx and add tests
Improve the structure of `src/CIAClassificationApp.tsx` and follow React best practices by modularizing components, adding tests, and setting up GitHub workflows.

* **Modularization**:
  - Add `src/components/Selection.tsx` and move the `Selection` component to this file.
  - Add `src/components/DetailCard.tsx` and move the `DetailCard` component to this file.
  - Add `src/hooks/useCIAOptions.ts` and move the options for availability, integrity, and confidentiality to this file.
  - Modify `src/CIAClassificationApp.tsx` to import the `Selection` and `DetailCard` components and the options from the new files.

* **Testing**:
  - Add `src/CIAClassificationApp.test.tsx` with unit tests for the `CIAClassificationApp` component using `@testing-library/react`.

* **GitHub Workflows**:
  - Add `.github/workflows/test-and-report.yml` to set up a GitHub workflow for running tests, generating coverage reports, and building the project.

* **Package Scripts**:
  - Modify `package.json` to include `test` and `coverage` scripts.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/Hack23/cia-compliance-manager?shareId=XXXX-XXXX-XXXX-XXXX).
