+++ src/CIAClassificationApp.tsx
@@ -1,224 +1,7 @@
 import React, { useState, useMemo } from "react";
-
-type CIADetails = {
-  capex: number;
-  opex: number;
-  impact: string;
-  technical: string;
-  description: string;
-};
-
-// Custom color mapping for each level, including "None"
-const levelColors: Record<string, { bg: string; text: string }> = {
-  None: { bg: "#e5e7eb", text: "#374151" },        // Grey tones for no controls
-  Basic: { bg: "#d1fae5", text: "#065f46" },         // Green hues
-  Moderate: { bg: "#fef3c7", text: "#92400e" },      // Yellow hues
-  High: { bg: "#ffedd5", text: "#c2410c" },          // Orange hues
-  "Very High": { bg: "#fee2e2", text: "#991b1b" },     // Red hues
-};
-
-// Option data for each CIA category with an added "None" option
-
-const availabilityOptions: Record<string, CIADetails> = {
-  None: {
-    capex: 0,
-    opex: 0,
-    impact: "No backup or recovery measures in place. Extremely high risk of downtime.",
-    technical: "No controls or redundancies implemented.",
-    description: "No Availability Controls: No backups or recovery options.",
-  },
-  Basic: {
-    capex: 5,
-    opex: 5,
-    impact: "Suitable for non-critical systems.",
-    technical:
-      "Manual backup and recovery with multiple SPOFs (Single Points of Failure).",
-    description: "Backup & Restore: Manual recovery, long RTO (~95% uptime).",
-  },
-  Moderate: {
-    capex: 15,
-    opex: 15,
-    impact:
-      "Works for mid-level critical systems, with faster recovery but some SPOFs remain.",
-    technical:
-      "Core systems are pre-configured but run only when needed; limited autoscaling.",
-    description:
-      "Pilot Light: Standby systems, automated recovery (~99% uptime).",
-  },
-  High: {
-    capex: 25,
-    opex: 40,
-    impact:
-      "Ideal for businesses with high uptime needs, such as online retailers.",
-    technical:
-      "Partial autoscaling and real-time data replication for faster failover.",
-    description: "Warm Standby: Fast recovery, limited SPOFs (~99.9% uptime).",
-  },
-  "Very High": {
-    capex: 60,
-    opex: 70,
-    impact:
-      "Necessary for mission-critical industries (e.g., finance, healthcare).",
-    technical:
-      "Full redundancy, autoscaling across regions, with instant recovery and zero downtime.",
-    description:
-      "Multi-Site Active/Active: Real-time failover (~99.99% uptime).",
-  },
-};
-
-const integrityOptions: Record<string, CIADetails> = {
-  None: {
-    capex: 0,
-    opex: 0,
-    impact: "No data validation or integrity measures in place. High risk for data corruption.",
-    technical: "No integrity or validation controls implemented.",
-    description: "No Integrity Controls: No validation or audit mechanisms.",
-  },
-  Basic: {
-    capex: 5,
-    opex: 10,
-    impact:
-      "Risk of data inaccuracies and compliance failures. Suitable for low-compliance businesses.",
-    technical: "No automated data checks; error-prone manual processes.",
-    description: "Manual Validation: Minimal checks, low auditability.",
-  },
-  Moderate: {
-    capex: 20,
-    opex: 20,
-    impact: "Meets basic compliance for industries like retail or general business.",
-    technical:
-      "Automated data checks and simple audit trails; moderate version control.",
-    description:
-      "Automated Validation: Enhanced accuracy and auditability.",
-  },
-  High: {
-    capex: 35,
-    opex: 50,
-    impact: "Ideal for highly regulated industries (finance, healthcare).",
-    technical:
-      "Blockchain used to ensure data integrity, ensuring data cannot be tampered with.",
-    description:
-      "Blockchain Validation: Immutable data records, high traceability.",
-  },
-  "Very High": {
-    capex: 60,
-    opex: 70,
-    impact:
-      "Perfect for industries needing full real-time data validation, like stock exchanges and defense contractors.",
-    technical:
-      "Smart contracts ensure real-time, automated execution of data-driven business rules.",
-    description:
-      "Smart Contracts: Real-time validation, full audit traceability.",
-  },
-};
-
-const confidentialityOptions: Record<string, CIADetails> = {
-  None: {
-    capex: 0,
-    opex: 0,
-    impact: "No encryption or access control. Data is fully exposed.",
-    technical: "No confidentiality measures implemented.",
-    description: "No Confidentiality Controls: No encryption or access restrictions.",
-  },
-  Basic: {
-    capex: 5,
-    opex: 5,
-    impact: "Suitable for public-facing data, no protection needed.",
-    technical: "Minimal security; HTTPS may be used for basic protection.",
-    description: "Public Data: No encryption or access control.",
-  },
-  Moderate: {
-    capex: 15,
-    opex: 20,
-    impact:
-      "Works for sensitive internal data (e.g., HR files, internal documents).",
-    technical:
-      "Strong encryption (AES-256), but with basic access control like password protection.",
-    description:
-      "Restricted Data: AES-256 encryption and basic monitoring.",
-  },
-  High: {
-    capex: 30,
-    opex: 40,
-    impact: "Essential for industries handling confidential data.",
-    technical: "MFA, robust encryption, continuous monitoring.",
-    description:
-      "Confidential Data: MFA, robust encryption, continuous monitoring.",
-  },
-  "Very High": {
-    capex: 60,
-    opex: 70,
-    impact: "Critical for top-secret or highly confidential data.",
-    technical:
-      "Advanced encryption protocols, strict access controls, and continuous security monitoring.",
-    description:
-      "Top Secret Data: Maximum protection with advanced protocols.",
-  },
-};
-
-// Reusable selection component with enhanced option labels and color styling
-interface SelectionProps {
-  label: string;
-  options: Record<string, CIADetails>;
-  value: string;
-  onChange: (value: string) => void;
-}
-
-const Selection: React.FC<SelectionProps> = ({ label, options, value, onChange }) => (
-  <div className="flex flex-col">
-    <label className="mb-2 font-medium text-gray-700 dark:text-gray-300">
-      {label}
-    </label>
-    <select
-      value={value}
-      onChange={(e) => onChange(e.target.value)}
-      className="p-2 border border-gray-300 dark:border-gray-700 rounded-md"
-    >
-      {Object.keys(options).map((level) => (
-        <option
-          key={level}
-          value={level}
-          style={{
-            backgroundColor: levelColors[level].bg,
-            color: levelColors[level].text,
-          }}
-        >
-          {level} - {options[level].description}
-        </option>
-      ))}
-    </select>
-    <p className="mt-2 text-sm" style={{ color: levelColors[value].text }}>
-      Impact: {options[value].impact}
-    </p>
-  </div>
-);
-
-// Reusable detail card component for displaying analysis with color accents
-interface DetailCardProps {
-  category: string;
-  level: string;
-  details: CIADetails;
-}
-
-const DetailCard: React.FC<DetailCardProps> = ({ category, level, details }) => (
-  <div
-    className="p-4 border border-gray-300 dark:border-gray-700 rounded-md"
-    style={{ backgroundColor: levelColors[level].bg }}
-  >
-    <h3 className="font-semibold" style={{ color: levelColors[level].text }}>
-      {category} - {level}
-    </h3>
-    <p className="text-sm text-gray-700 dark:text-gray-200">
-      CAPEX: {details.capex}% | OPEX: {details.opex}%
-    </p>
-    <p className="mt-1" style={{ color: levelColors[level].text }}>
-      Impact: {details.impact}
-    </p>
-    <p className="mt-1 text-gray-600 dark:text-gray-400">
-      Technical: {details.technical}
-    </p>
-  </div>
-);
+import Selection from "./components/Selection";
+import DetailCard from "./components/DetailCard";
+import { availabilityOptions, integrityOptions, confidentialityOptions } from "./hooks/useCIAOptions";
 
 const CIAClassificationApp: React.FC = () => {
   const [availability, setAvailability] = useState<string>("None");
@@ -338,4 +121,3 @@ const CIAClassificationApp: React.FC = () => {
 };
 
 export default CIAClassificationApp;
-
